+++ pricing/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from pricing.factory import get_provider
+from pricing import ui
 
-__all__ = ['get_provider']
+__all__ = ['get_provider', 'ui']